test: move tests into roles
Put tests into a `test/` directory in each role so that they're closer
to the stuff they're testing. This will also make it easier to split a
role from this repo.

In the process, convert tests to be actual playbooks, bringing the tests
closer to what will be used in production.  This also makes it easier to
run the test independently of the test framework and makes for some nice
examples.

Based on a patch by Leos Pol <lpol@redhat.com>.